+++ create-react-ank.js
@@ -1,6 +1,10 @@
 #!/usr/bin/env node
 
 const { spawnSync } = require('child_process');
+const readline = require('readline').createInterface({
+  input: process.stdin,
+  output: process.stdout
+});
 
 const repositoryUrl = 'https://github.com/ankithg03/react-18-boilerplate';
 
@@ -24,6 +28,38 @@ const cloneProcess = spawnSync('git', ['clone', repositoryUrl, destinationDirect
 
 if (cloneProcess.status === 0) {
   console.log('Repository cloned successfully.');
+
+  const packageJsonPath = `${destinationDirectory}/package.json`;
+
+  const fs = require('fs');
+
+  try {
+    const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'));
+
+    const updatePackageJson = (field, defaultValue) => {
+      const existingValue = packageJson[field] || defaultValue;
+      readline.question(`Enter a value for ${field} in package.json (or press Enter to keep the existing value: [${existingValue}]): `, (input) => {
+        if (input.trim() !== '') {
+          packageJson[field] = input;
+        } else {
+          packageJson[field] = defaultValue;
+        }
+
+        if (field === 'repository') {
+          fs.writeFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2));
+          console.log('Package.json updated with user-provided values.');
+          readline.close();
+        } else {
+          updatePackageJson('name', packageJson.name);
+        }
+      });
+    };
+
+    updatePackageJson('name', packageJson.name);
+  } catch (error) {
+    console.error('Error reading or updating package.json:', error);
+  }
 } else {
   console.error('Error cloning repository.');
+  process.exit(1);
 }