+++ CHANGELOG.md
@@ -18,3 +18,11 @@ and this project is to [CREATE A REACT APP](https://www.npmjs.com/package/create
 #### Changed
     - Added support for cloning the repository to a specified directory.
     - Added a help option to display usage information.
+
+### [1.0.2] - 2023-11-07
+
+#### Added
+    - Added interactive prompts for updating package.json fields.
+    - Supported keeping existing values in package.json if no input provided by the user.
+    - Enhanced user prompts by showing existing data in brackets for each field.
+